refactor: Seperate subreddit resources from bot resource manager
